requirements.txt
streamlit
opencv-python
deepface
pillow
numpy
